+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "circuits/circuits/packages/efficient-zk-sig"]
+	path = circuits/circuits/packages/efficient-zk-sig
+	url = https://github.com/personaelabs/efficient-zk-sig
